requirements.txt
numpy>=1.26,<3
pandas>=2.2,<3
requests>=2.31,<3
scikit-learn>=1.4,<2
snowflake-connector-python>=3.7,<4
pytest>=8,<9
